Generation of DB model from sqlite system tables, missing tests for sqlite, a little bit of renames

+++ src/DotNetDBTools.DefinitionParser/Agnostic/AgnosticDefaultValueMapper.cs
@@ -20,7 +20,7 @@ namespace DotNetDBTools.DefinitionParser.Agnostic
             return dataType switch
             {
                 StringDataType stringDataType => (string)value,
-                IntDataType intDataType => (int)value,
+                IntDataType intDataType => (long)(int)value,
                 ByteDataType byteDataType => (byte[])value,
                 _ => throw new InvalidOperationException($"Invalid default value type: '{value.GetType()}' for a column with type '{dataType.GetType()}'"),
             };
