+++ src/DotNetDBTools.DefinitionParser/MSSQL/MSSQLDefaultValueMapper.cs
@@ -23,7 +23,7 @@ namespace DotNetDBTools.DefinitionParser.MSSQL
             return dataType switch
             {
                 StringDataType stringDataType => (string)value,
-                IntDataType intDataType => (int)value,
+                IntDataType intDataType => (long)(int)value,
                 ByteDataType byteDataType => (byte[])value,
                 IUserDefinedType userDefinedType => MapByColumnDataType(userDefinedType.UnderlyingType, value),
                 _ => throw new InvalidOperationException($"Invalid default value type: '{value.GetType()}' for a column with type '{dataType.GetType()}'"),
